chore: move `scripts/src/*` to `scripts/*`
