+++ .husky/commit-msg
@@ -1,4 +1,4 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
 
-npx --no-install esno scripts/src/verifyCommit.ts $1
+npx --no-install esno scripts/verifyCommit.ts $1
