--- scripts/README.md
@@ -1,3 +1,0 @@
-# umi-scripts
-
-See our website [umijs](https://umijs.org) for more information.
