+++ scripts/bundleDeps.ts
