+++ scripts/checkPackageFiles.ts
@@ -2,7 +2,7 @@ import { glob, lodash, logger } from '@umijs/utils';
 import { isMatch } from 'matcher';
 import 'zx/globals';
 import { PATHS } from './.internal/constants';
-import { eachPkg, getPkgs } from './utils';
+import { eachPkg, getPkgs } from './.internal/utils';
 
 const COMMON_IGNORES = [
   // default included
