+++ scripts/jest-turbo.ts
@@ -1,5 +1,5 @@
 import { PATHS } from './.internal/constants';
-import { spawnSync } from './utils';
+import { spawnSync } from './.internal/utils';
 
 (async () => {
   spawnSync(`jest -c ${PATHS.JEST_TURBO_CONFIG}`, { cwd: process.cwd() });
