+++ scripts/release.ts
@@ -5,7 +5,7 @@ import { join } from 'path';
 import rimraf from 'rimraf';
 import 'zx/globals';
 import { PATHS } from './.internal/constants';
-import { assert, eachPkg, getPkgs } from './utils';
+import { assert, eachPkg, getPkgs } from './.internal/utils';
 
 (async () => {
   const { branch } = getGitRepoInfo();
