+++ scripts/setupWebStorm.ts
@@ -1,5 +1,5 @@
 import { PATHS } from './.internal/constants';
-import { eachPkg, getPkgs, setExcludeFolder } from './utils';
+import { eachPkg, getPkgs, setExcludeFolder } from './.internal/utils';
 
 const cwd = process.cwd();
 eachPkg(getPkgs(), ({ name }) => {
