+++ scripts/verifyCommit.ts
